move to extensions/autoDeleteTrans.js
